Optimize incremental compilation in case of MarkRange.h changes
The problem is that Chunk.h is very common header, and it includes
MarkRange, so let's simply move ChunkInfoCloneable<MarkRangesInfo> into
MarkRange.h
