+++ src/Processors/Chunk.h
@@ -3,7 +3,6 @@
 #include <Columns/IColumn_fwd.h>
 #include <Common/CollectionOfDerived.h>
 #include <Core/Types_fwd.h>
-#include <Storages/MergeTree/MarkRange.h>
 
 #include <memory>
 
@@ -157,30 +156,6 @@ public:
 };
 
 using AsyncInsertInfoPtr = std::shared_ptr<AsyncInsertInfo>;
-
-/// Lineage information: from which table, part and mark range does the chunk come from?
-/// This information is needed by the query condition cache.
-class MarkRangesInfo : public ChunkInfoCloneable<MarkRangesInfo>
-{
-public:
-    MarkRangesInfo(UUID table_uuid_, const String & part_name_, size_t marks_count_, bool has_final_mark_, MarkRanges mark_ranges_)
-        : table_uuid(table_uuid_)
-        , part_name(part_name_)
-        , marks_count(marks_count_)
-        , has_final_mark(has_final_mark_)
-        , mark_ranges(std::move(mark_ranges_))
-    {}
-
-    void appendMarkRanges(const MarkRanges & mark_ranges_);
-
-    UUID table_uuid;
-    String part_name;
-    size_t marks_count;
-    bool has_final_mark;
-    MarkRanges mark_ranges;
-};
-
-using MarkRangesInfoPtr = std::shared_ptr<MarkRangesInfo>;
 
 /// Converts all columns to full serialization in chunk.
 /// It's needed, when you have to access to the internals of the column,
