+++ src/Storages/MergeTree/MarkRange.cpp
@@ -3,6 +3,7 @@
 #include <IO/ReadHelpers.h>
 #include <IO/WriteHelpers.h>
 
+#include <base/defines.h>
 #include <fmt/ranges.h>
 
 namespace DB
@@ -11,6 +12,11 @@ namespace DB
 namespace ErrorCodes
 {
     extern const int LOGICAL_ERROR;
+}
+
+MarkRange::MarkRange(size_t begin_, size_t end_) : begin(begin_), end(end_)
+{
+    chassert(begin <= end);
 }
 
 size_t MarkRange::getNumberOfMarks() const
@@ -115,6 +121,18 @@ void MarkRanges::deserialize(ReadBuffer & in)
         readBinaryLittleEndian((*this)[i].begin, in);
         readBinaryLittleEndian((*this)[i].end, in);
     }
+}
+
+MarkRangesInfo::MarkRangesInfo(UUID table_uuid_, const String & part_name_, size_t marks_count_, bool has_final_mark_, MarkRanges mark_ranges_)
+    : table_uuid(table_uuid_)
+    , part_name(part_name_)
+    , marks_count(marks_count_)
+    , has_final_mark(has_final_mark_)
+    , mark_ranges(mark_ranges_)
+{}
+void MarkRangesInfo::appendMarkRanges(const MarkRanges & mark_ranges_)
+{
+    mark_ranges.insert(mark_ranges.end(), mark_ranges_.begin(), mark_ranges_.end());
 }
 
 }
